+++ apps/web/app/components/blocks/structure/MultiGrid/types.ts
@@ -75,3 +75,10 @@ export type ColumnBlock = Block & {
 };
 
 export type GapSize = 'None' | 'S' | 'M' | 'L' | 'XL';
+
+export interface VisibleGridState {
+  filteredToOriginal: number[];
+  originalToFiltered: Record<number, number>;
+  columnWidths: number[];
+  blocks: Block[];
+}
